# coinfetcher/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Coin Fetcher API",
        "version":"version 1.0" 
    },
    "paths": {
        "/v1/health": {
            "get": {
                "description": "This gets the Gecko API health status",
                "produces": [
                    "application/json"
                ],
                "summary": "Get Gecko API health status Endpoint",
                "operationId": "checkHealth",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/price_api.HealthResponse"
                        }
                    }
                }
            }
        },
        "/v1/price": {
            "get": {
                "description": "Fetches the price of a given coin ticker.",
                "produces": [
                    "application/json"
                ],
                "summary": "Fetch coin price Endpoint",
                "operationId": "fetchPrice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Coin ticker symbol",
                        "name": "ticker",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/price_api.PriceResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "price_api.HealthResponse": {
            "type": "object",
            "properties": {
                "geckoapistatus": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "price_api.PriceResponse": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "number"
                },
                "ticker": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "vol24Hr": {
                    "type": "number"
                }
            }
        }
    }
}